extract kyc response
