+++ services/regulated-assets-approval-server/internal/serve/kyc-status/post_handler.go
@@ -27,30 +27,6 @@ func (h PostHandler) validate() error {
 	return nil
 }
 
-type postRequest struct {
-	CallbackID   string
-	EmailAddress string `json:"email_address"`
-}
-
-type postResponse struct {
-	Result  string `json:"result"`
-	Message string `json:"message"`
-}
-
-func NewApprovedKYCStatusPostResponse() *postResponse {
-	return &postResponse{
-		Result:  "no_further_action_required",
-		Message: "Your KYC has been approved!",
-	}
-}
-
-func NewRejectedKYCStatusPostResponse() *postResponse {
-	return &postResponse{
-		Result:  "no_further_action_required",
-		Message: "Your KYC has been rejected!",
-	}
-}
-
 func (h PostHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	err := h.validate()
